fix: address pr review issues
- backend: broaden exception handling in _get_default_error_message from
  Project.DoesNotExist only to Exception, covering Django ValidationError
  raised on malformed UUID strings

- backend: add 5-minute cache for api_error_message lookup to avoid a DB
  hit on every error path; sentinel pattern caches both hits (custom msg)
  and misses (None) so projects without a custom message also benefit

- views: move ProjectApiErrorMessageView from nexus/inline_agents/api/views
  to nexus/projects/api/views where project settings belong

- views: fetch project before input validation in patch() so a
  nonexistent project_uuid returns 404 instead of 400

- views: remove 500-character limit for api_error_message

- views: use Project.objects.only() to load only needed fields instead
  of full object via get_project_by_uuid

- views: invalidate api_error_message cache key on patch save

- router: update ProjectApiErrorMessageView import to new location

- tests: update notify_async mock paths to nexus.projects.api.views and
  remove test that covered the now-deleted 500-char length limit
