+++ nexus/inline_agents/api/views.py
@@ -1536,50 +1536,3 @@ class ProjectEngineSourceView(APIView):
         return Response(data={"engine_source": "OWN" if has_own else "STANDARD"})
 
 
-API_ERROR_MESSAGE_MAX_LENGTH = 500
-
-
-class ProjectApiErrorMessageView(APIView):
-    permission_classes = [IsAuthenticated, ProjectPermission]
-
-    def get(self, request, project_uuid):
-        try:
-            project = get_project_by_uuid(project_uuid)
-        except ProjectDoesNotExist:
-            return Response(data={"error": "Project not found"}, status=404)
-
-        return Response(data={"error_message": project.api_error_message})
-
-    def patch(self, request, project_uuid):
-        if "error_message" not in request.data:
-            return Response(data={"error": "error_message is required"}, status=400)
-
-        error_message = request.data.get("error_message")
-        if error_message is not None and not isinstance(error_message, str):
-            return Response(data={"error": "error_message must be a string"}, status=400)
-
-        normalized_message = error_message.strip() if isinstance(error_message, str) else None
-        if normalized_message == "":
-            normalized_message = None
-
-        if normalized_message and len(normalized_message) > API_ERROR_MESSAGE_MAX_LENGTH:
-            return Response(
-                data={"error": f"error_message must be at most {API_ERROR_MESSAGE_MAX_LENGTH} characters"},
-                status=400,
-            )
-
-        try:
-            project = get_project_by_uuid(project_uuid)
-        except ProjectDoesNotExist:
-            return Response(data={"error": "Project not found"}, status=404)
-
-        project.api_error_message = normalized_message
-        project.save(update_fields=["api_error_message"])
-        notify_async(event="cache_invalidation:project", project=project)
-
-        return Response(
-            data={
-                "message": "Error message updated successfully",
-                "error_message": project.api_error_message,
-            }
-        )
